DIP_lab_1_BT23CSE182/pyproject.toml
[project]
name = "dip-lab-1-bt23cse182"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.5",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
]
